commit 3

+++ src/main/java/peaksoft/dto/response/RestaurantResponse.java
@@ -6,7 +6,7 @@ public record RestaurantResponse(Long id,
                                  String name,
                                  String location,
                                  String restType,
-                                 int service) {
+                                 double service) {
 
 
 }
